+++ Assets/Dust/Scripts/Editor/Gizmos/ConeGizmoEditor.cs
@@ -3,22 +3,21 @@ using UnityEditor;
 
 namespace DustEngine.DustEditor
 {
-    [CustomEditor(typeof(DuPyramidGizmo))]
+    [CustomEditor(typeof(ConeGizmo))]
     [CanEditMultipleObjects]
-    public class DuPyramidGizmoEditor : DuGizmoObjectEditor
+    public class ConeGizmoEditor : GizmoObjectEditor
     {
         private DuProperty m_Radius;
         private DuProperty m_Height;
-        private DuProperty m_Faces;
         private DuProperty m_Center;
         private DuProperty m_Direction;
 
         //--------------------------------------------------------------------------------------------------------------
 
-        [MenuItem("Dust/Gizmos/Pyramid")]
+        [MenuItem("Dust/Gizmos/Cone")]
         public static void AddComponentToSelectedObjects()
         {
-            AddGizmoToSelectedOrNewObject(typeof(DuPyramidGizmo));
+            AddGizmoToSelectedOrNewObject(typeof(ConeGizmo));
         }
 
         //--------------------------------------------------------------------------------------------------------------
@@ -29,7 +28,6 @@ namespace DustEngine.DustEditor
 
             m_Radius = FindProperty("m_Radius", "Radius");
             m_Height = FindProperty("m_Height", "Height");
-            m_Faces = FindProperty("m_Faces", "Faces");
             m_Center = FindProperty("m_Center", "Center");
             m_Direction = FindProperty("m_Direction", "Direction");
         }
@@ -42,18 +40,13 @@ namespace DustEngine.DustEditor
 
             PropertyExtendedSlider(m_Radius, 0f, 10f, 0.01f);
             PropertyExtendedSlider(m_Height, 0f, 10f, 0.01f);
-            PropertyExtendedIntSlider(m_Faces, 3, 64, 1, 3, 256);
             PropertyField(m_Center);
             PropertyField(m_Direction);
             Space();
             PropertyField(m_Color);
             PropertyField(m_GizmoVisibility);
 
-            // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-            // Validate & Normalize Data
-
-            if (m_Faces.isChanged)
-                m_Faces.valInt = DuPyramidGizmo.Normalizer.Faces(m_Faces.valInt);
+            // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
             InspectorCommitUpdates();
         }