+++ Assets/Dust/Scripts/Runtime/Fields/2D/TextureSpaceField.cs
@@ -275,7 +275,7 @@ namespace DustEngine
             Gizmos.matrix = transform.localToWorldMatrix;
 
             Gizmos.color = GetGizmoColorRange1();
-            DuGizmos.DrawRect(width, height, Vector3.zero, direction);
+            DustGizmos.DrawRect(width, height, Vector3.zero, direction);
         }
 #endif