+++ Assets/Dust/Scripts/Runtime/Fields/2D/WaveField.cs
@@ -268,7 +268,7 @@ namespace DustEngine
                 zeroPoint.x = GetPowerForLocalPositionInAxisXPlus(new Vector3(0f, linearFalloff, linearFalloff)); // this point will be always outside falloff range
                 zeroPoint = DuAxisDirection.ConvertFromAxisXPlusToDirection(direction, zeroPoint);
 
-                DuGizmos.DrawCircle(linearFalloff, zeroPoint, direction, 32);
+                DustGizmos.DrawCircle(linearFalloff, zeroPoint, direction, 32);
             }
         }
 #endif