+++ Assets/Dust/Scripts/Runtime/Gizmos/TorusGizmo.cs
@@ -2,11 +2,11 @@
 
 namespace DustEngine
 {
-    [AddComponentMenu("Dust/Gizmos/Cone Gizmo")]
-    public class DuConeGizmo : DuGizmoObject
+    [AddComponentMenu("Dust/Gizmos/Torus Gizmo")]
+    public class TorusGizmo : GizmoObject
     {
         [SerializeField]
-        private float m_Radius = 1f;
+        private float m_Radius = 2f;
         public float radius
         {
             get => m_Radius;
@@ -14,11 +14,11 @@ namespace DustEngine
         }
 
         [SerializeField]
-        private float m_Height = 2f;
-        public float height
+        private float m_Thickness = 0.5f;
+        public float thickness
         {
-            get => m_Height;
-            set => m_Height = value;
+            get => m_Thickness;
+            set => m_Thickness = value;
         }
 
         [SerializeField]
@@ -30,8 +30,8 @@ namespace DustEngine
         }
 
         [SerializeField]
-        private Axis6xDirection m_Direction = Axis6xDirection.YPlus;
-        public Axis6xDirection direction
+        private Axis3xDirection m_Direction = Axis3xDirection.Y;
+        public Axis3xDirection direction
         {
             get => m_Direction;
             set => m_Direction = value;
@@ -41,7 +41,7 @@ namespace DustEngine
 
         public override string GizmoName()
         {
-            return "Cone";
+            return "Torus";
         }
 
 #if UNITY_EDITOR
@@ -50,7 +50,7 @@ namespace DustEngine
             Gizmos.matrix = transform.localToWorldMatrix;
             Gizmos.color = color;
 
-            DuGizmos.DrawWireCone(radius, height, center, direction, 64, 4);
+            DustGizmos.DrawWireTorus(radius, thickness, center, direction, 64, 32);
         }
 #endif
     }